31 | Comment 1:N relationship

+++ articles/views.py
@@ -1,10 +1,11 @@
 from IPython import embed
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponseForbidden
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views.decorators.http import require_POST, require_GET
-# from accounts.models import User
 from django.contrib.auth import get_user_model
+
 from .forms import ArticleForm, CommentForm
 from .models import Article, Comment
 
@@ -62,11 +63,11 @@ def detail(request, article_pk):
 @require_POST
 def delete(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    # if request.method == 'POST':
-    article.delete()
-    return redirect('articles:index')
-    # else:
-    #     return redirect('articles:detail', article.pk)
+    if article.user == request.user:
+        article.delete()
+        return redirect('articles:index')
+    else:
+        return HttpResponseForbidden()
 
 # def edit(request, article_pk):
 #     article = Article.objects.get(pk=article_pk)
@@ -74,22 +75,25 @@ def delete(request, article_pk):
 #         'article': article
 #     }
 #     return render(request, 'articles/edit.html', context)
-
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        article_form = ArticleForm(request.POST, instance=article)
-        if article_form.is_valid():
-            article = article_form.save()
-            return redirect('articles:detail', article.pk)
+    if article.user == request.user:
+        if request.method == 'POST':
+            article_form = ArticleForm(request.POST, instance=article)
+            if article_form.is_valid():
+                article = article_form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            article_form = ArticleForm(instance=article)
+        context = {
+            'article_form': article_form
+        }
+        return render(request, 'articles/form.html', context)
     else:
-        article_form = ArticleForm(instance=article)
-    context = {
-        'article_form': article_form
-    }
-    return render(request, 'articles/form.html', context)
+        return HttpResponseForbidden()
 
 @require_POST
+@login_required
 def comment_create(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
     # 1. modelform에 사용자 입력값 넣고
@@ -101,6 +105,7 @@ def comment_create(request, article_pk):
         comment = comment_form.save(commit=False)
         # 3-2. FK 넣고 저장
         comment.article = article
+        comment.user = request.user
         comment.save()
     # 4. return redirect
     else:
@@ -116,7 +121,10 @@ def comment_create(request, article_pk):
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    # messages.add_message(request, messages.INFO, '댓글이 삭제 되었습니다.')
-    messages.success(request, '댓글이 삭제되었습니다.')
-    return redirect('articles:detail', article_pk)
+    if request.user == comment.user:
+        comment.delete()
+        # messages.add_message(request, messages.INFO, '댓글이 삭제 되었습니다.')
+        messages.success(request, '댓글이 삭제되었습니다.')
+        return redirect('articles:detail', article_pk)
+    else:
+        return HttpResponseForbidden()
